Fix refactor leftovers
