+++ src/renderer/hooks/useDragAndDrop.js
@@ -1,0 +1,52 @@
+import { useState, useCallback } from "react";
+
+function useDragAndDrop(onFilesDropped) {
+  const [isDragging, setIsDragging] = useState(false);
+
+  const handleDragEnter = useCallback((e) => {
+    e.preventDefault();
+    e.stopPropagation();
+    setIsDragging(true);
+  }, []);
+
+  const handleDragLeave = useCallback((e) => {
+    e.preventDefault();
+    e.stopPropagation();
+    if (e.currentTarget.contains(e.relatedTarget)) return;
+    setIsDragging(false);
+  }, []);
+
+  const handleDragOver = useCallback((e) => {
+    e.preventDefault();
+    e.stopPropagation();
+  }, []);
+
+  const handleDrop = useCallback((e) => {
+    e.preventDefault();
+    e.stopPropagation();
+    setIsDragging(false);
+
+    const files = Array.from(e.dataTransfer.files);
+    if (files.length > 0 && onFilesDropped) {
+      const fileObjects = files.map((file) => ({
+        path: file.path || file.name,
+        name: file.name,
+        type: 'file',
+        size: file.size
+      }));
+      onFilesDropped(fileObjects);
+    }
+  }, [onFilesDropped]);
+
+  return {
+    isDragging,
+    dragProps: {
+      onDragEnter: handleDragEnter,
+      onDragLeave: handleDragLeave,
+      onDragOver: handleDragOver,
+      onDrop: handleDrop
+    }
+  };
+}
+
+export default useDragAndDrop;
